requirements.txt
Django==5.0.1
Pillow>=10.0.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
python-decouple==3.8
django-extensions==3.2.3
gunicorn
whitenoise